Using BigDecimal to deal with money.

+++ src/main/scala/Apple.scala
@@ -1,7 +1,7 @@
 package com.jm2dev.hmrc
 
-class Apple(quantity: Int) extends Fruit(quantity, 0.6f) {
-  override val cost: Float = new Fruit(offeredQuantity(), 0.6f).cost
+class Apple(quantity: Int) extends Fruit(quantity, BigDecimal(0.6)) {
+  override val cost: BigDecimal = new Fruit(offeredQuantity(), BigDecimal(0.6)).cost
 
   private def offeredQuantity() = Math.floorMod(quantity, 2) match {
     case 0 => quantity / 2
