+++ src/main/scala/Fruit.scala
@@ -1,6 +1,6 @@
 package com.jm2dev.hmrc
 
-class Fruit(quantity: Int, price: Float) {
+class Fruit(quantity: Int, price: BigDecimal) {
   val cost = quantity * price
 
   def total() = "£%.2f" format cost
